% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullpower.R
\name{bullpower}
\alias{bullpower}
\title{bull power technical analysis function}
\usage{
bullpower(h,top)
}
\arguments{
\item{h}{an stock data}

\item{top}{an rsi rise horizon value}
}
\value{
an analysis of stock data for bull power technical analysis indicators
}
\description{
bull power technical analysis function is to analyze the reversal pattern conform to the rising trend and bull market of stock data
}
\details{
use RSI analysis of the strength of the stock market trend, analyze trends conform to bull power, and RSI function need library 'TTR'
}
\examples{
library(quantmod)
aapl<-getSymbols("AAPL",src="google",auto.assign=FALSE)
bullpower(aapl,60)
}
\author{
Chun-Yu Liu <john401528@gmail.com>
}
