% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_function.R
\name{down_function}
\alias{down_function}
\title{down trend line down function}
\usage{
down_function(h,day,num)
}
\arguments{
\item{h}{an stock data}

\item{day}{the number of days to analysis the data}

\item{num}{select pivot calculation method
1:  UP1<-(2*center)-Lo(h).
2:  UP2<-center+(UP1-DOWN1).}
}
\value{
an analysis of stock data for down function
}
\description{
Down trend line analysis of the down_function is to sort and analysis the stock data
}
\details{
down_function is to sort the stock data by the set number of days and to filter the data of rising trend
}
\examples{
library(quantmod)
aapl<-getSymbols("AAPL",src="yahoo",auto.assign=FALSE)
down_function(aapl,20,1)
}
\author{
Chun-Yu Liu <john401528@gmail.com>
}
