% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downtrend.R
\name{downtrend}
\alias{downtrend}
\title{downtrend technical analysis function}
\usage{
downtrend(h,day,num)
}
\arguments{
\item{h}{an stock data}

\item{day}{the number of days to analysis the data}

\item{num}{select pivot calculation method
1:  UP1<-(2*center)-Lo(h).
2:  UP2<-center+(UP1-DOWN1).}
}
\value{
an analysis of stock data for down trend technical analysis indicators
}
\description{
Down trend line technical analysis function is to analyze the downward trend of stock data
}
\details{
use the down_function analysis data to analysis the downward trend line
}
\examples{
library(quantmod)
aapl<-getSymbols("AAPL",src="yahoo",auto.assign=FALSE)
downtrend(aapl,20,1)
}
\author{
Chun-Yu Liu <john401528@gmail.com>
}
