% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/up_function.R
\name{up_function}
\alias{up_function}
\title{up trend line up function}
\usage{
up_function(h,day,num)
}
\arguments{
\item{h}{an stock data}

\item{day}{the number of days to analysis the data}

\item{num}{select pivot calculation method
1:  DOWN1<-(2*center)-Hi(h).
2:  DOWN2<-center-(UP1-DOWN1).}
}
\value{
an analysis of stock data for up function
}
\description{
Up trend line analysis of the up_function is to sort and analysis the stock data
}
\details{
up_function is to sort the stock data by the set number of days and to filter the data of downward trend
}
\examples{
library(quantmod)
aapl<-getSymbols("AAPL",src="google",auto.assign=FALSE)
up_function(aapl,20,1)
}
\author{
Chun-Yu Liu <john401528@gmail.com>
}
