% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uptrend.R
\name{uptrend}
\alias{uptrend}
\title{uptrend technical analysis function}
\usage{
uptrend(h,day,num)
}
\arguments{
\item{h}{an stock data}

\item{day}{the number of days to analysis the data}

\item{num}{select pivot calculation method
1:  DOWN1<-(2*center)-Hi(h).
2:  DOWN2<-center-(UP1-DOWN1).}
}
\value{
an analysis of stock data for up trend technical analysis indicators
}
\description{
Up trend line technical analysis function is to analyze the rising trend of stock data
}
\details{
use the up_function analysis data to analysis the rising trend line
}
\examples{
\dontrun{
library(quantmod)
aapl<-getSymbols("AAPL",src="yahoo",auto.assign=FALSE)
uptrend(aapl,20,1)
}
}
\author{
Chun-Yu Liu <john401528@gmail.com>
}
