% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_bottom.R
\name{v_bottom}
\alias{v_bottom}
\title{v bottom technical analysis function}
\usage{
v_bottom(h,top,down,month,day)
}
\arguments{
\item{h}{an stock data}

\item{top}{an rsi rise horizon value}

\item{down}{an rsi down horizon value}

\item{month}{set the length between the start and end points. Unit:month}

\item{day}{check the correctness of the end point, set the length between the end and check points. Unit:day}
}
\value{
an analysis of stock data for v bottom technical analysis indicators
}
\description{
V bottom technical analysis function is to analyze the rising trend of stock data
}
\details{
use RSI analysis of the strength of the stock market trend, analyze trends conform to v bottom, and RSI function need library 'TTR'
}
\examples{
\dontrun{
library(quantmod)
aapl<-getSymbols("AAPL",src="yahoo",auto.assign=FALSE)
v_bottom(aapl,60,40,3,20)
}
}
\author{
Chun-Yu Liu <john401528@gmail.com>
}
