% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_top.R
\name{v_top}
\alias{v_top}
\title{v top technical analysis function}
\usage{
v_top(h,top,down,month,day)
}
\arguments{
\item{h}{an stock data}

\item{top}{an rsi rise horizon value}

\item{down}{an rsi down horizon value}

\item{month}{set the length between the start and end points. Unit:month}

\item{day}{check the correctness of the end point, set the length between the end and check points. Unit:day}
}
\value{
an analysis of stock data for v top technical analysis indicators
}
\description{
V top technical analysis function is to analyze the downward trend of stock data
}
\details{
use RSI analysis of the strength of the stock market trend, analyze trends conform to v top, and RSI function need library 'TTR'
}
\examples{
\dontrun{
library(quantmod)
aapl<-getSymbols("AAPL",src="google",auto.assign=FALSE)
v_top(aapl,60,40,3,20)
}
}
\author{
Chun-Yu Liu <john401528@gmail.com>
}
